+++ src/imio/helpers/patches.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from imio.helpers.cache import invalidate_cachekey_volatile_for
 from Products.PlonePAS.plugins.role import GroupAwareRoleManager
+from Products.PluggableAuthService.plugins.ZODBRoleManager import ZODBRoleManager
 from smtplib import SMTP_SSL
 from zope.sendmail.mailer import SMTPMailer
 
@@ -38,7 +39,13 @@ def assignRolesToPrincipal(self, roles, principal_id, REQUEST=None):  # noqa
     invalidate_cachekey_volatile_for('_users_groups_value')
 
 
-def assignRoleToPrincipal(self, role_id, principal_id, REQUEST=None):
-    GroupAwareRoleManager._old_assignRoleToPrincipal(self, role_id, principal_id, REQUEST)
+def assignRoleToPrincipal(self, role_id, principal_id):
+    ZODBRoleManager._old_assignRoleToPrincipal(self, role_id, principal_id)
+    # we need to invalidate cachekey
+    invalidate_cachekey_volatile_for('_users_groups_value')
+
+
+def removeRoleFromPrincipal(self, role_id, principal_id):
+    ZODBRoleManager._old_removeRoleFromPrincipal(self, role_id, principal_id)
     # we need to invalidate cachekey
     invalidate_cachekey_volatile_for('_users_groups_value')